// src/api/model/updateStatus.ts
/**
 * Generated by orval v7.0.1 🍺
 * Do not edit manually.
 * Paratrooper API
 * OpenAPI spec version: 1.0.0
 */

export type UpdateStatus = (typeof UpdateStatus)[keyof typeof UpdateStatus];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const UpdateStatus = {
  pending: "pending",
  processing: "processing",
  published: "published",
  failed: "failed",
  canceled: "canceled",
} as const;