// src/api/model/update.ts
/**
 * Generated by orval v7.0.1 🍺
 * Do not edit manually.
 * Paratrooper API
 * OpenAPI spec version: 1.0.0
 */
import type { UpdateStatus } from "./updateStatus";

export type Update = {
  channel: string;
  createdAt: string;
  id: string;
  message: string;
  runtimeVersion: string;
  status: UpdateStatus;
};